% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cldtable.R
\name{cldtable}
\alias{cldtable}
\title{Compact Letter Display Table}
\usage{
cldtable(modelx, emmeans, round_digits = 3, stat_adjust = "Tukey")
}
\arguments{
\item{modelx}{A fitted model object}

\item{emmeans}{An emmeans object}

\item{round_digits}{Number of digits to round to (default 3)}

\item{stat_adjust}{P-value adjustment method (default "Tukey")}
}
\value{
Data frame with CLD annotations
}
\description{
Create compact letter display tables from model and emmeans
}
