% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_large_text.R
\name{clean_large_text}
\alias{clean_large_text}
\title{Clean Large Text}
\usage{
clean_large_text(x)
}
\arguments{
\item{x}{responses to clean}
}
\value{
response vector that is preprocessed
}
\description{
This function allows you to preprocess text responses with more than 2 words.
}
