% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_naresponses.R
\name{clean_naresponses}
\alias{clean_naresponses}
\title{Clean na responses}
\usage{
clean_naresponses(word)
}
\arguments{
\item{word}{word to clean, if multiple use loop/apply/dplyr}
}
\value{
responses changed to NA when indicating lack of knowledge
}
\description{
This function allows you to change responses that indicate lack of knowledge into NA.
}
