% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_sent.R
\name{Code_sent}
\alias{Code_sent}
\title{Code single-word responses into dictionaries and sentiment}
\usage{
Code_sent(data, raw_text = "rawword", preproc_text = "word")
}
\arguments{
\item{data}{dataframe including the text to code}

\item{raw_text}{string name of a variable (vector) of text that has not been preprocessed}

\item{preproc_text}{string name of a variable (vector) of text that has been preprocessed using Full_preprocess()}
}
\value{
Original dataframe with additional columns for codings
}
\description{
This function allows you to code a variable of text responses into different sentiment scores. If available, sentiments are matched to raw words, otherwise they are matched to preprocessed words.Average the different valence scores. If a SWN score is available for the correct sense, then average that with other valence scores, if not, then check if version one of the valence scores means is na, if so, get the average of the second versions, if not get the average of the first versions.
}
