% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_single.R
\name{Code_single}
\alias{Code_single}
\title{Code single-word responses}
\usage{
Code_single(data, text = "word", more2na = T)
}
\arguments{
\item{data}{dataframe including the text to code}

\item{text}{string name of a variable (vector) of text to code into the dictionaries}

\item{more2na}{if text has more than 2 words, transform to na}
}
\value{
Original dataframe with additional columns for codings
}
\description{
This function allows you to code a variable of (fully preprocessed, see "preprocess_words") text responses into the different dictionaries
}
