% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohens_d_EMmeans.R
\name{Cohens_d_EMmeans}
\alias{Cohens_d_EMmeans}
\title{Cohen's d from Emmeans}
\usage{
Cohens_d_EMmeans(mod, pairwises)
}
\arguments{
\item{mod}{A fitted mixed model object}

\item{pairwises}{Pairwise comparisons from emmeans}
}
\value{
Data frame with contrast info and Cohen's d values
}
\description{
Compute Cohen's d from emmeans pairwise comparisons
}
