% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dictionaries.R
\name{compute_socats_directions}
\alias{compute_socats_directions}
\title{Compute SOCATS-specific direction indicators}
\usage{
compute_socats_directions(data)
}
\arguments{
\item{data}{Data frame with SOCATS binary columns}
}
\value{
Data frame with direction columns added
}
\description{
Compute SOCATS-specific direction indicators
}
\keyword{internal}
