% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinize.R
\name{cosinize}
\alias{cosinize}
\title{Cosine Similarity Computation}
\usage{
cosinize(vector, dictionaries)
}
\arguments{
\item{vector}{Numeric vector}

\item{dictionaries}{Matrix where each column is a reference vector}
}
\value{
Named numeric vector of cosine similarities
}
\description{
Compute cosine similarity between a vector and columns of a matrix
}
