% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_ending_Ss.R
\name{delete_ending_Ss}
\alias{delete_ending_Ss}
\title{Singular transform}
\usage{
delete_ending_Ss(word)
}
\arguments{
\item{word}{word to transform, if multiple use loop/apply/dplyr}
}
\value{
word vector without the last trailing S
}
\description{
This function deletes the letter s from the input whenever it trails the word
}
