% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_internals.R
\name{.has_java_home}
\alias{.has_java_home}
\title{Check whether a directory looks like a Java home}
\usage{
.has_java_home(path)
}
\arguments{
\item{path}{Candidate Java home directory}
}
\value{
Logical scalar
}
\description{
Check whether a directory looks like a Java home
}
\keyword{internal}
