% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_internals.R
\name{.has_namespace}
\alias{.has_namespace}
\title{Check namespace availability}
\usage{
.has_namespace(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Logical scalar
}
\description{
Check namespace availability
}
\keyword{internal}
