% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_internals.R
\name{.reticulate_import}
\alias{.reticulate_import}
\title{Wrapper around reticulate::import for testability}
\usage{
.reticulate_import(module)
}
\arguments{
\item{module}{Python module name}
}
\value{
Imported Python module
}
\description{
Wrapper around reticulate::import for testability
}
\keyword{internal}
