% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_internals.R
\name{.reticulate_py_to_r}
\alias{.reticulate_py_to_r}
\title{Wrapper around reticulate::py_to_r for testability}
\usage{
.reticulate_py_to_r(x)
}
\arguments{
\item{x}{Python object}
}
\value{
Converted R object
}
\description{
Wrapper around reticulate::py_to_r for testability
}
\keyword{internal}
