% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_internals.R
\name{.sleep_seconds}
\alias{.sleep_seconds}
\title{Wrapper around Sys.sleep for testability}
\usage{
.sleep_seconds(seconds)
}
\arguments{
\item{seconds}{Numeric sleep duration}
}
\value{
NULL invisibly
}
\description{
Wrapper around Sys.sleep for testability
}
\keyword{internal}
