% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanize.R
\name{euclideanize}
\alias{euclideanize}
\title{Euclidean Distance Computation}
\usage{
euclideanize(vector, dictionaries)
}
\arguments{
\item{vector}{Numeric vector}

\item{dictionaries}{Matrix where each column is a reference vector}
}
\value{
Named numeric vector of Euclidean distances
}
\description{
Compute Euclidean distance between a vector and columns of a matrix
}
