% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Antonyms.R
\name{get_Antonyms}
\alias{get_Antonyms}
\title{Wordnet antonym retriever}
\usage{
get_Antonyms(synsets, Syns = TRUE)
}
\arguments{
\item{synsets}{synsets to obtain antonyms for. May be a single synset or a
list of synsets.}

\item{Syns}{Return the synsets? Defaults to TRUE. If FALSE, returns the
various words that make up the synsets.}
}
\value{
antonym expansion
}
\description{
Gets antonyms for nouns or adjectives of a term's synset.
}
