% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Syns.R
\name{get_Syns}
\alias{get_Syns}
\title{Wordnet Synset retriever}
\usage{
get_Syns(term, PoS = "NOUN", sense = NULL, Syns = T)
}
\arguments{
\item{term}{word to obtain the synset for}

\item{PoS}{Part of speech of the word. Defaults to "NOUN". Other options are "VERB", and "ADJECTIVE"}

\item{sense}{sense of the term for which the synset is desired. Defaults to NULL.}

\item{Syns}{Return the synset? Defaults to TRUE. If FALSE, returns the various words that make up the synset.}
}
\value{
wordnet synsets
}
\description{
Provides the Wordnet synset for a word.
}
