% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Obituary_data}
\alias{Obituary_data}
\title{Example obituary data for text coding}
\format{
A data frame with obituary text responses
}
\usage{
Obituary_data
}
\description{
Example obituary data for use with the text coding pipeline
}
\keyword{datasets}
