% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_internals.R
\name{replace_nan_with_na}
\alias{replace_nan_with_na}
\title{NaN-aware replacement for data frames}
\usage{
replace_nan_with_na(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Data frame with NaN values replaced by NA
}
\description{
NaN-aware replacement for data frames
}
\keyword{internal}
