% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SADCAT-package}
\alias{SADCAT}
\alias{SADCAT-package}
\title{SADCAT package}
\description{
Dictionary creation with stereotype content dictionaries.
}
\author{
\strong{Maintainer}: Gandalf Nicolas \email{gandalfn@gmail.com}

}
\keyword{internal}
