% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sentiments}
\alias{Sentiments}
\title{Sentiment lexicon data}
\format{
A data frame with sentiment words and their valence scores
}
\usage{
Sentiments
}
\description{
Combined sentiment lexicon data used for valence scoring
}
\keyword{datasets}
