% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_preprocess.R
\name{Full_preprocess}
\alias{Full_preprocess}
\title{Full preprocess}
\usage{
Full_preprocess(words, parallelize = T, print = T, debug = F)
}
\arguments{
\item{words}{text to preprocess}

\item{parallelize}{use parallel processors?}
}
\value{
fully preprocessed words
}
\description{
This function applies all response preprocessing functions included in the package. (includes lowering, trimming whitespace, eliminating spaces and symbols, lemmatizing, and spellcheking; some of these only work on single words)
}
