% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Hypos.R
\name{get_Hypos}
\alias{get_Hypos}
\title{Wordnet hyponym retriever}
\usage{
get_Hypos(synsets, Syns = T)
}
\arguments{
\item{synsets}{synsets to obtain hyponyms for}

\item{Syns}{Return the synsets? Defaults to TRUE. If FALSE, returns the various words that make up the synsets.}
}
\value{
hyponym hierarchy
}
\description{
Gets the synsets/words (depending on Syns) of a term's hyponym hierarchy.
}
