% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Member_holonym.R
\name{get_Member_holonym}
\alias{get_Member_holonym}
\title{Wordnet member holonym retriever}
\usage{
get_Member_holonym(synsets, Syns = T)
}
\arguments{
\item{synsets}{synsets to obtain member holonyms  for}

\item{Syns}{Return the synsets? Defaults to TRUE. If FALSE, returns the various words that make up the synsets.}
}
\value{
member holonyms of the words
}
\description{
Gets member holonyms for a term's synset
}
