% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_part_meronyms.R
\name{get_part_meronyms}
\alias{get_part_meronyms}
\title{Wordnet part meronyms retriever}
\usage{
get_part_meronyms(synsets, Syns = T)
}
\arguments{
\item{synsets}{synsets to obtain part meronyms for}

\item{Syns}{Return the synsets? Defaults to TRUE. If FALSE, returns the various words that make up the synsets.}
}
\value{
part meronyms of the words
}
\description{
Gets part meronyms for a term's synset
}
