% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seed_Vectors_allwords_Avg}
\alias{Seed_Vectors_allwords_Avg}
\title{Average word embeddings for seed dictionaries (all words)}
\format{
A data frame with seed vectors computed from all words
}
\usage{
Seed_Vectors_allwords_Avg
}
\description{
Average word embeddings for seed dictionaries using all words
}
\keyword{datasets}
