% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seed_Vectors_personcontext_Avg}
\alias{Seed_Vectors_personcontext_Avg}
\title{Average word embeddings for seed dictionaries (person context)}
\format{
A data frame with seed vectors computed from person context words
}
\usage{
Seed_Vectors_personcontext_Avg
}
\description{
Average word embeddings for seed dictionaries using person context words
}
\keyword{datasets}
